pyproject.toml
[tool.poetry]
name = "flask-ses-rq"
version = "0.1.0"
description = ""
authors = ["AgathaBahati <matinodev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^3.0.3"
psycopg2-binary = "^2.9.9"
flask-testing = "^0.8.1"
flask-sqlalchemy = "^3.1.1"
flask-wtf = "^1.2.1"
email-validator = "^2.2.0"
python-dotenv = "^1.0.1"
bootstrap-flask = "^2.4.0"
redis = "^5.0.7"
rq = "^1.16.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
